Add script to automate installation
I'm a big fan of making potential contributors do as little manual work
as possible to set up the app. I've updated the installation instructions
to add more details about the prerequisites and how to install them, and
added a bootstrap script that takes care of installing elasticsearch,
running it in the background, installing the gem dependencie, and
importing the sample data.