Removing redundant calls to create_index
Created a DataMagic::Config.index_exists? method to ensure that the
index only gets created once. If the index exists after the app starts
subsequent calls to  #import_with_dictionary and #import_csv
will not trigger index creation.

On the other hand, if the index is not present on app start, then
those actions will create the index.

Working on issue #135