Fixed server error with NULL fields
When restricting search results to a certain set of fields, ES does not
by default include fields with missing or NULL values. It can
happen that the only field searched for is a  missing value field, which
threw a nil server error. This is now fixed by handling the error
and also re-inserting the searched-for fields with a nil value.

Fixes Issue #186