Extracts an output object from indexing
        Hey, so this extraction is a little odd. It doesn't have the
one place only log flow and it feels like this is a product issue. Maybe
we should normalized logging to get a dump of data per row cycle instead
of all over the place.

        Also, seems to me like there is a missing repository type object
in here, that creates (client.index) or updates (client.update) and maybe
we don't need to know which it is doing out at this level.

        Another possibility is that the doc_id logic moves to a doc
decorator and there might also be a place in there for this create or
update stuff.